projects/dn_flow/requirements.txt
pandas
pandas-gbq
pyarrow
oauth2client
feedparser
google-auth
azure-identity
azure-keyvault-secrets
prefect